+++ irsim/env/env_plot.py
@@ -174,7 +174,7 @@ class EnvPlot:
             label (str): Label for the trajectory.
             show_direction (bool): Whether to show the direction of the trajectory.
             refresh (bool): Whether to refresh the plot.
-            kwargs: Additional plotting options.
+            kwargs: Additional plotting options for ax.plot()
         """
         if isinstance(traj, list):
             path_x_list = [p[0, 0] for p in traj]
@@ -289,9 +289,9 @@ class EnvPlot:
             keep_len (int): Length of the last frame. Default is 30.
             rm_fig_path (bool): Whether to remove the figure path after saving. Default is True.
             kwargs: Additional arguments for saving the animation.
-                See https://imageio.readthedocs.io/en/v2.8.0/format_gif-pil.html for details.
+                See `format_gif <https://imageio.readthedocs.io/en/v2.8.0/format_gif-pil.html>`_ for details.
         """
-        self.saved_ani_kwargs.update({"subrectangles": True})
+        self.saved_ani_kwargs.update({"subrectangles": True, "loop": 0})
         self.saved_ani_kwargs.update(kwargs)
 
         env_param.logger.info("Start to create animation")