+++ irsim/usage/14world_3d_plot/robot_world_3d.py
@@ -1,0 +1,14 @@
+import irsim
+
+# env = irsim.make('car_world.yaml', projection='3d')
+env = irsim.make('robot_world_3d.yaml', projection='3d')
+
+for i in range(1000):
+
+    env.step()
+    env.render(0.05)
+
+    if env.done():
+        break
+
+env.end()