+++ irsim/util/util.py
@@ -515,7 +515,7 @@ def random_point_range(range_low=[0, 0, -pi], range_high=[10, 10, pi]):
     if isinstance(range_low, list):
         range_low = np.c_[range_low]
 
-    if isinstance(range_low, list):
+    if isinstance(range_high, list):
         range_high = np.c_[range_high] 
 
     return np.random.uniform(range_low, range_high)